Refactor: 시간 데이터 문자열로 변환하는 부분 유틸로 추출
